# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "WChat 即时通讯系统后端 API 文档",
        "title": "WChat API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/groups": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建一个新的群聊，当前用户自动成为群主；可选择在创建时邀请初始成员",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "群组"
                ],
                "summary": "创建群聊",
                "parameters": [
                    {
                        "description": "群组信息",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.CreateGroupReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "参数错误",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/groups/joined": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取当前用户已加入的所有群聊，包含群名、成员数、群主等信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "群组"
                ],
                "summary": "获取我加入的群列表",
                "responses": {
                    "200": {
                        "description": "Token 无效",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/groups/{uuid}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据群 UUID 获取群名称、公告、头像、成员数、群主、加群模式等信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "群组"
                ],
                "summary": "获取群详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "群 UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "群不存在",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "群主修改群名称、公告、头像、加群模式等（仅传入需修改的字段）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "群组"
                ],
                "summary": "修改群资料",
                "parameters": [
                    {
                        "type": "string",
                        "description": "群 UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "需要更新的字段",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.UpdateGroupReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "权限不足 / 群不存在",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "群主解散群聊（仅群主有权限），解散后群内所有会话失效",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "群组"
                ],
                "summary": "解散群聊",
                "parameters": [
                    {
                        "type": "string",
                        "description": "群 UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "权限不足 / 群不存在",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/groups/{uuid}/members": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取指定群聊的所有成员，包含昵称、头像、签名等信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "群组"
                ],
                "summary": "获取群成员列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "群 UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "群不存在",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "邀请一个或多个用户加入指定群聊",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "群组"
                ],
                "summary": "邀请成员入群",
                "parameters": [
                    {
                        "type": "string",
                        "description": "群 UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "被邀请的用户 UUID 列表",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.InviteMemberReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "参数错误 / 群不存在",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/groups/{uuid}/members/me": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "当前用户主动退出指定群聊（群主不能退出，需先转让或解散）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "群组"
                ],
                "summary": "退出群聊",
                "parameters": [
                    {
                        "type": "string",
                        "description": "群 UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "不在群中 / 群主不能退出",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/groups/{uuid}/members/{user_id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "群主/管理员将指定用户从群聊中移除（不能踢自己）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "群组"
                ],
                "summary": "踢出群成员",
                "parameters": [
                    {
                        "type": "string",
                        "description": "群 UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "被踢用户 UUID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "权限不足 / 不能踢自己",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "types.CreateGroupReq": {
            "type": "object",
            "required": [
                "members",
                "name"
            ],
            "properties": {
                "add_mode": {
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ]
                },
                "avatar": {
                    "type": "string"
                },
                "members": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string",
                    "maxLength": 20
                },
                "notice": {
                    "type": "string",
                    "maxLength": 500
                }
            }
        },
...
        "types.InviteMemberReq": {
            "type": "object",
            "required": [
                "user_ids"
            ],
            "properties": {
                "user_ids": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "types.UpdateGroupReq": {
            "type": "object",
            "properties": {
                "add_mode": {
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ]
                },
                "avatar": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 20
                },
                "notice": {
                    "type": "string",
                    "maxLength": 500
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}